Merge pull request #1 from sp98/fix-data-flow

+++ app/src/main/java/com/example/cowinalert/CreateAlertViewModel.kt
@@ -74,7 +74,7 @@ class CreateAlertViewModel(
         // create instance of alert data class
         val alert = Alert(
             name = name,
-            pinCode = pin,
+            pinCode = pin.toLong(),
             isCovishield = isCovishield,
             isCovaxin = isCovaxin,
             above45 = isAbove45,
